pyproject.toml
[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[project]
name = "com-mgmtp-unbounddb"
version = "0.1.0"
description = "Helps with Unbound"
authors = [
  {name = "Jonas Voss", email = "jonas.voss@mgm-tp.com"}
]
maintainers = [
  {name = "Jonas Voss", email = "jonas.voss@mgm-tp.com"}
]
readme = "README.md"
classifiers = [
  # check https://pypi.org/classifiers/ for further information
  "Development Status :: 1 - Planning",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  # optinally specify supported python versions here
  "Programming Language :: Python :: 3.11",
  "Typing :: Typed",
  "Operating System :: Unix",
  "Operating System :: MacOS",
]
requires-python = ">=3.11,<4"
dependencies = [
  "pyyaml",
  "pydantic",
  "pydantic-settings",
  "duckdb>=1.0.0",
  "polars>=1.0.0",
  "pyarrow>=15.0.0",
  "httpx>=0.27.0",
  "typer>=0.12.0",
  "streamlit>=1.37.0",
]

[project.scripts]
unbounddb = "unbounddb.cli:app"

[dependency-groups]
local = [
  "bump-my-version",
  "pre-commit",
]
dev = [
  { include-group = "lint" },
  { include-group = "test" },
]
lint = [
  "ruff",
  "mypy",
  "types-pyyaml",
]
test = [
  "pytest",
  "coverage",
]
doc = [
  "sphinx",
  "autodoc-pydantic>=2.2.0",
  "pydata-sphinx-theme",
  "sphinx-copybutton",
  "sphinx-design",
]


[project.urls]
Repository = "empty"



[[tool.uv.index]]
name = "mgmart"
url = "https://artifacts.mgm-tp.com/artifactory/api/pypi/pypi-repos/simple"

[tool.uv.build-backend]
module-name = "unbounddb"
module-root = ""

[tool.mypy]
# see https://mypy.readthedocs.io/en/stable/config_file.html
# you can also disable certain warnings inline via "# type: ignore" comment
# Import discovery
mypy_path = "."
files = ["./unbounddb"]
plugins = 'pydantic.mypy'
namespace_packages = false
explicit_package_bases = false
ignore_missing_imports = true
follow_imports = "normal"
follow_imports_for_stubs = false
no_site_packages = false
# Platform setttings
python_version = "3.11"
# Disallow dynamic typing
disallow_any_unimported = false
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_generics = true
disallow_subclassing_any = false
# Untyped definitions and calls
disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
# None and Optional handling
no_implicit_optional = true
strict_optional = true
# Configuring warnings
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
# Suppressing errors
ignore_errors = false
# Miscellaneous strictness flags
allow_untyped_globals = false
allow_redefinition = false
local_partial_types = false
implicit_reexport = true
strict_equality = true
# Configuring error messages
show_error_context = false
show_column_numbers = false
show_error_codes = true
pretty = true

[tool.pytest.ini_options]
addopts = ["--strict-config", "--strict-markers"]
testpaths = ["tests/unittests"]

[tool.coverage.run]
source = ["unbounddb"]
relative_files = true
branch = true
omit = [
  '*/__init__.py',  # __init__.py files should normally not contain any testable code
  '*/__main__.py'   # same applies to __main__.py files (they should only be entry points)
]

[tool.coverage.report]
skip_empty = true
# you can also omit lines for the coverage report via the "# pragma: no cover" comment
exclude_lines = [
  'pragma: no cover',
  '@abstract',
  '@overload',
  'raise AssertionError',
  'raise NotImplementedError',
  '\.\.\.',
  'if __name__ == .__main__.:',
  'if TYPE_CHECKING',
  'if typing.TYPE_CHECKING',
  'pass'
]
